+++ l10n_br_nfe/models/res_partner.py
@@ -128,3 +128,12 @@ class ResPartner(spec_models.SpecModel):
         for rec in self:
             if rec.nfe40_fone:
                 rec.phone = rec.nfe40_fone
+
+    def _export_field(self, xsd_field, class_obj, member_spec):
+        if xsd_field == 'nfe40_xNome' and \
+                class_obj._name == 'nfe.40.dest' and \
+                self.env.context.get('tpAmb') == '2':
+            return 'NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO' \
+                   ' - SEM VALOR FISCAL'
+        return super(ResPartner, self)._export_field(
+            xsd_field, class_obj, member_spec)
